updated resources and added get all users
